telegram-bot/requirements.txt
pycryptodome
aiogram==3.22.0
python-dotenv
sqlalchemy
loguru
python-jose
aiosqlite
openpyxl
pillow